+++ web/src/contexts/FontSizeContext.ts
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+export interface FontSizeContextType {
+  isFontSizeLarge: boolean;
+  toggleFontSize: () => void;
+}
+
+export const FontSizeContext = createContext<FontSizeContextType>({
+  isFontSizeLarge: false,
+  toggleFontSize: () => {},
+});
